refactor: reorganize utilities into domain-specific files
Replaced expectUtils.ts and utils.ts with domain-specific files:

- formatting.ts: formatPosition (can later add formatBlock, etc.)
- inventory.ts: getInventorySummary, expectItemInInventory, equipItem, expectSufficientItems
- entities.ts: getEntityName, expectEntity
- blocks.ts: isBlockEmpty, expectBlock, expectBlockOfType

This organization groups functions by what they operate on rather than by
function type, making the codebase more intuitive to navigate.
